Change version of Org to Long from Integer for HBase Dao compatibility

+++ smart-user-api/src/main/java/com/smartitengineering/user/domain/Organization.java
@@ -13,7 +13,7 @@ import org.codehaus.jackson.annotate.JsonIgnore;
  *
  * @author russel
  */
-public class Organization extends AbstractGenericPersistentDTO<Organization, String, Integer> {
+public class Organization extends AbstractGenericPersistentDTO<Organization, String, Long> {
 
   private String name;
   private Address address;
